wip: ike-sa: Add property for interface ID
wip: Should we maintain the if-ids during mbb reauth? (Currently not
possible as there is no setter.)  Maybe not and handling of the CHILD_SAs
might have to change too. We currently copy if-ids of CHILD_SAs to the new
child-create jobs, however, this might not actually be necessary/useful.
Unlike during rekeying these are really new SAs, so updown scripts are called
and using the same if-ids might actually confuse these.